pyproject.toml
[project]
name = "beyond-chat-user-persona-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.25.0",
    "praw>=7.8.1",
]
